+++ ProyectoDesarrollo/src/Model/Manager.java
@@ -13,7 +13,7 @@ public class Manager extends User{
     
     public Manager(String id, String nombre, String cedula, 
             String cargo, String correo, int genero, String direccion, 
-            String telefono, double salario, String fechaNacimiento, 
+            String telefono, float salario, String fechaNacimiento, 
             String cuentaBancaria, String fechaRegistro, String nombreUsuario,
             String contrasena, boolean habilitado) {
         super(id, nombre, cedula, cargo, correo,
